// src/generated-types/menu-item.ts
export interface MenuItemTranslation {
  id: string;
  title: string;
  description?: string | null | undefined;
  language: string;
}
// ...
export interface CreateMenuItemRequest {
  slug: string;
  price: string;
  imageUrl?: string | null | undefined;
  isAvailable?: boolean | null | undefined;
  categoryId: string;
}

/** Request message for updating a menu item. */
export interface UpdateMenuItemRequest {
  id: string;
  slug?: string | null | undefined;
  price?: string | null | undefined;
  imageUrl?: string | null | undefined;
  isAvailable?: boolean | null | undefined;
}

/** Request message for changing the position of a menu item. */
export interface ChangeMenuItemPositionRequest {
  id: string;
  position: number;
}

/** Request message for creating a menu item translation. */
export interface CreateMenuItemTranslationRequest {
  title: string;
  description?: string | null | undefined;
  language: string;
  itemId: string;
}

/** Request message for updating a menu item translation. */
export interface UpdateMenuItemTranslationRequest {
  id: string;
  title?: string | null | undefined;
  description?: string | null | undefined;
}